cleans names
