+++ packages/core/src/components/ColorInput/example.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-alert */
 import * as BUI from "../..";
 
-BUI.Manager.registerComponents();
+BUI.Manager.init();
 
 const colorInput =
   document.body.querySelector<BUI.ColorInput>("bim-color-input")!;
