+++ packages/core/src/components/Selector/example.html
@@ -17,11 +17,11 @@
 </head>
 
 <body>
-  <bim-selector-input label="Camera Projection">
+  <bim-selector label="Camera Projection">
     <bim-option icon="solar:settings-bold" label="Orthographic" value="0"></bim-option>
     <bim-option label="Perspective" value="1"></bim-option>
     <bim-option label="Custom"></bim-option>
-  </bim-selector-input>
+  </bim-selector>
   <bim-tabs tab="controls">
     <bim-tab name="controls" label="Controls"></bim-tab>
     <bim-tab name="" label=""></bim-tab>
