+++ packages/core/src/components/Toolbar/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./Group";
+export * from "./Section";
+export * from "./Toolbar";
